fix: resolve all compilation errors - BUILD SUCCESSFUL
- Add google() to pluginManagement repositories in settings.gradle.kts
- Remove kotlin version from module plugins (provided by root)
- Add local.properties with SDK path
- Add suppressUnsupportedCompileSdk=35 to gradle.properties
- Fix Compose compiler version to 1.5.4 (compatible with Kotlin 1.9.20)
- Fix signing config val reassignment (use different variable names)
- Add DocumentRef data class to SupremeOntology.kt
- Fix device-abstraction module: add universal-model dependency, fix MutableSharedFlow, Instant, val reassignment, suspend calls
- Fix FixAIEngine: wrap CombinedAnalysis before passing to methods
- Fix HomeHubEngine: add CommandResult sealed class, nullable Boolean
- Fix MaintenanceOSEngine: smart cast workarounds with local vals, !! for nullable
- Fix VibrationDoctorEngine: Float to Double conversion
- Fix NoiseDoctorEngine: Float to Double conversion
- Fix LeakWatchEngine: nullable flowRateLpm handling
- Fix WarrantyVaultEngine: Int to Long conversion
- Fix BaseViewModel: use getApplication() instead of private application
- Fix NoiseViewModel: match NoiseAnalysis field names (noiseType, diagnosis.causes)
- Fix VibrationViewModel: match VibrationAnalysis field names (rmsG, baselineDeviation.rmsDeviation)
- Remove document-scanner (use play-services-mlkit-document-scanner instead)

BUILD SUCCESSFUL - APK: 78 MB

+++ supreme/modules/noise-doctor/src/main/kotlin/com/supreme/noise/NoiseDoctorEngine.kt
@@ -227,7 +227,7 @@ class NoiseDoctorEngine {
 
         return SpectralFeatures(
             centroid = centroid,
-            rolloff = rolloff,
+            rolloff = rolloff.toDouble(),
             bandwidth = bandwidth,
             rms = rms,
             peak = peak,
